\name{blur-method}
\alias{blur}
\alias{blur-method}
\alias{blur,ANY-method}
\docType{methods}
\title{Method "blur"}
\description{
  This method could be thought of as a ...
}
\usage{
blur(object, x, \dots)
}
\arguments{
  \item{object}{object of implementing class}
  \item{x}{TBD}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\keyword{methods}

