\name{Umpire-package}
\alias{Umpire-package}
\alias{Umpire}
\docType{package}
\title{UMPIRE: Ultimate Microarray Prediction, Inference, and Reality Engine}
\description{
  A suite of microarray simulation software which includes additive and
  multiplicative noise, mixture of expressed and unexpressed genes, and
  uses statistical distributions to capture differences in mean expression
  and in standard deviation both within groups and between groups of samples.
  Finally, it incorporates a simple block correlation structure between genes.
}
\details{
  \tabular{ll}{
    Package: \tab Umpire\cr
    Type: \tab Package\cr
    Version: \tab 1.2.3\cr
    Date: \tab 2011-12-01\cr
    License: \tab Artistic-2.0\cr
    LazyLoad: \tab yes\cr
  }

  For a complete list of functions, use \code{library(help = 'Umpire')}.
}
\references{OOMPA}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  Jiexin Zhang \email{jiexinzhang@mdanderson.org},
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

