\name{EngineWithActivity-class}
\alias{EngineWithActivity-class}
\alias{EngineWithActivity}
\alias{alterMean,EngineWithActivity-method}
\alias{alterSD,EngineWithActivity-method}
\alias{rand,EngineWithActivity-method}
\alias{nrow,EngineWithActivity-method}
\alias{summary,EngineWithActivity-method}

\title{The EngineWithActivity Class}
\description{The \code{EngineWithActivity} is used to set some
  components in the object of class \code{Engine} to be
  transcriptionally inactive and transform the expression data to
  appropriate logarithmic scale.
  
}
\usage{
EngineWithActivity(active, components, base = 2)
\S4method{alterMean}{EngineWithActivity}(object, TRANSFORM, \dots)
\S4method{alterSD}{EngineWithActivity}(object, TRANSFORM, \dots)
\S4method{rand}{EngineWithActivity}(object,n,\dots)
\S4method{summary}{EngineWithActivity}(object, \dots)
\S4method{nrow}{EngineWithActivity}(x)
}
\arguments{
  \item{active}{Object of class \code{logical} with length equaling to the
    number of components specifying whether each component should be
    transcriptionally active or not. Or, \code{active} can be a numeric
    within the range of 0 to 1 specifying the probability for a component to
    be active.
  }
  \item{components}{Object of class \code{list}, each element of which
    contains the parameters for the underlying distribution that the
    gene expression follows.}
  \item{base}{A numeric specifying the logarithmic scale to which the data should
    be transformed.
  }
  \item{object, x}{Object of class \code{EngineWithActivity}}
  \item{TRANSFORM}{the \code{TRANSFORM} function for each object should
    take as its input a vector of mean expression or standard deviation
    and return a transformed vector that can be used to alter the
    appropriate slot of the object.}
  \item{n}{Number of samples to be simulated}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\section{Objects from the Class}{
Objects can be created by the use of the \code{EngineWithActivity}
generator function.
}
\section{Slots}{
  \describe{
    \item{\code{active}:}{Either an object of class \code{logical}
      specifying whether each component should be transcriptionally
      active or not, or a numeric specifying the probability for a
      component to be transcriptionally active}
    \item{\code{base}:}{The logarithmic scale.}
    \item{\code{components}:}{Object of class \code{"list"} specifying
      the parameters of the underlying distribution.}
  }
}
\section{Details}{
  An ENGINE WITH ACTIVITY allows for the possibility that some
  components (or genes) in an expression engine (or tissue) might be
  transcriptionally inactive.  Thus, the true biological signal
  S_gi should really be viewed as a mixture:
  
  S_gi = z_g * delta_0 + (1 - z_g) * T_gi

  where delta_0 = a point mass at zero;   
  T_gi = a random variable supported on the positive real line;   
  z_g ~ Binom(pi) defines the activity state (1 = on, 0 = off)

  The 'rand' method for an EngineWithActivity is a little bit
  tricky, since we do two things at once. First, we use the
  'base' slot to exponentiate the random variables generated by
  the underlying Engine on the log scale.  We treat base = 0 as
  a special case, which means that we should continue to work on
  the scale of the Engine.  Second, we mask any inactive component
  by replacing the generated values with 0.

  Note that this is terribly inefficient if we only have a single
  homogeneous population, since we generate a certain amount of
  data only to throw it away.  The power comes when we allow
  cancer disregulation to turn a block on or off, when the
  underlying data reappears.
}

\section{Methods}{
  \describe{
    \item{alterMean(object, TRANSFORM,\dots)}{takes an object of class
      \code{EngineWithActivity},loop over the components and alter the
      mean as defined by TRANSFORM function, and returns a object of class
      \code{EngineWithActivity} with modified components and the same
      active and base slots}
    \item{alterSD(object, TRANSFORM,\dots)}{takes an object of class
      \code{EngineWithActivity},loop over the components and alter the
      standard deviation as defined by TRANSFORM function, and returns a
      modified object of class \code{EngineWithActivity} with modified
      components and the same active and base slots}
    \item{rand(object,n,\dots)}{generates nrow(EngineWithActivity)*n matrix
      representing gene expressions of n samples, and the transcriptionally
      inactive components are masked by 0}
    \item{summary(object,\dots)}{prints out the total number of
      components and the number of active components in the object
      of \code{EngineWithActivity}}
    \item{nrow(x)}{counts the total number of genes (i.e, the
      length of the vector the \code{EngineWithActivity} will generate)}
  }
}

\value{
  The \code{EngineWithActivity} generator returns an object of class
  \code{EngineWithActivity} with slots described in \code{Slots}
  section.
  
  The \code{alterMean} returns an object of class \code{EngineWithActivity} with
  altered mean
  
  The \code{alterSD} returns an object of class \code{EngineWithActivity} with
  altered standard deviation
  
  The \code{rand} returns nrow(EngineWithActivity)*n gene expression matrix with the
  inactive components being masked by 0.

  The \code{summary} prints out the total number of
  components and the number of active components in the object
  of \code{EngineWithActivity}
  
  The \code{nrow} returns the number of genes (i.e, the length of the
  vector) the \code{EngineWithActivity} object will generate.
}
\references{KRC}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  Jiexin Zhang \email{jiexinzhang@mdanderson.org}
}

\seealso{
  \code{\linkS4class{Engine}}
}
\examples{
nComponents <- 10
nGenes <- 100
active <- 0.7
comp <- list()
for (i in 1:nComponents){
  comp[[i]] <- IndependentNormal(rnorm(nGenes/nComponents,6,1.5),1/rgamma(nGenes/nComponents,44,28))
}
myEngine <- EngineWithActivity(active,comp,2)
summary(myEngine)
myData <- rand(myEngine,5)
dim(myData)
}
\keyword{datagen}
\keyword{classes}
