\name{Umpire-package}
\alias{Umpire-package}
\alias{Umpire}
\docType{package}
\title{
UMPIRE: Ultimate Microarray Prediction, Inference, and Reality Engine
}
\description{
A suite microarray simulation software which includes additive and
multiplicative noise, mixture of expressed and unexpressed genes, and
uses statistical distributions to capture differences in mean expression and in
standard deviation both within groups and between groups of samples. Finally, we incorporate
a simple block correlation structure between genes.
}
\details{
\tabular{ll}{
Package: \tab Umpire\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-13\cr
License: \tab Perl Artistic License (see LICENSE file)\cr
LazyLoad: \tab yes\cr
}
For a complete list of functions, use \code{library(help = 'Umpire')}.
}
\references{KRC}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  Jiexin Zhang \email{jiexinzhang@mdanderson.org}
}
\keyword{ package }

