\name{summary-method}
\alias{summary}
\alias{summary-method}
\alias{summary, ANY-method}

\title{Method "summary"}
\description{
  Print out summary information for object. Content being printed out
  depends on the object passed to the method.
}
\usage{
summary(object, ...)
}
\arguments{
  \item{object}{object of class \code{Engine}}
  \item{\dots}{extra arguments for generic or plotting routines}
}
\references{KRC}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  Jiexin Zhang \email{jiexinzhang@mdanderson.org}
}